Move the mobile companion RPC facade into its own extension file
The companion pairing actions and the full CompanionRequest dispatcher move
to AppModel+MobileCompanion.swift. The two fileprivate entry points the
gateway wiring calls become internal, and the core members the dispatcher
reads (backend, taskWorkers, workspaceAccess, queued-run and schedule
dispatch helpers) widen to internal with an 'internal(for: AppModel
extension files)' breadcrumb — the same seam every later facade extension
split relies on. Chat-export entry points stay in the class body for now;
they ride with the settings/export facade split.

Behavior-identical refactor; project regenerated with xcodegen.
